requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
brotli==1.0.9
    # via flask-compress
build==0.8.0
    # via pip-tools
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   flask
    #   pip-tools
dash==2.6.2
    # via -r requirements/requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
flask==2.2.2
    # via
    #   dash
    #   flask-compress
flask-compress==1.13
    # via dash
fredapi==0.5.0
    # via -r requirements/requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
numpy==1.23.4
    # via pandas
packaging==21.3
    # via build
pandas==1.5.1
    # via
    #   -r requirements/requirements.in
    #   fredapi
pep517==0.13.0
    # via build
pip-tools==6.9.0
    # via -r requirements/requirements.in
plotly==5.10.0
    # via dash
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.21.0
    # via -r requirements/requirements.in
pytz==2022.5
    # via pandas
requests==2.28.1
    # via -r requirements/requirements.in
six==1.16.0
    # via python-dateutil
tenacity==8.1.0
    # via plotly
tomli==2.0.1
    # via
    #   build
    #   pep517
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via flask
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
